requirements.txt
attrs==24.2.0
autoflake==1.6.1
bidict==0.23.1
black==22.8.0
blinker==1.8.2
bokeh==3.0.1
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
filelock==3.15.4
Flask==3.0.3
Flask-SocketIO==5.3.6
fonttools==4.53.1
h11==0.14.0
identify==2.6.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.2
more-itertools==10.4.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==2.20.0
py==1.11.0
pyflakes==3.2.0
pyparsing==3.1.2
pytesseract==0.3.10
pytest==7.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
simple-websocket==1.0.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
watchdog==4.0.2
Werkzeug==3.0.3
wsproto==1.2.0
xyzservices==2024.6.0
